Simplify CommonStmt in parse tree
Original merge commit: 5f5430af333a46e843cc7fdefcb9b759ec1bb642

######

Simplify CommonStmt in parse tree
The grammar requires parsing the first common block in a common stmt
differently from the others: the '//' is optional for the blank common.

But once it's parsed, it is easier to work with if each is represented
as a `parser::CommonStmt::Block`. This is achieved by using the same
constructor for `CommonStmt` but then including the first block in
the list of blocks.


Original commit: dd46afd6b5607ec61ea9b2e97efa16deda31441b
-------