Remove ObjectName alternative from Designator
Original merge commit: 7c56fc3afbbd6dfa065f0659d7eee2cbb5638a1c

######

Remove ObjectName alternative from Designator
A simple name in a `Designator` is always parsed as a `DataRef`, not
an `ObjectName`. So remove that alternative.

`StmtFunctionStmt::ConvertToAssignment` was creating a `Designator` with
that alternative: change it to do the same thing as the parser.

Add `GetSimpleName` utility functions to check if an `Expr` or `Variable`
represents a simple name. Many of the places that checked for `ObjectName`
in `Designator` are trying to do that.

Clean up includes and forward declarations in `tools.h`.


Original commit: 97d44de7b146aee601caa0bc529f7af8a6be001e
-------