Distinguish between unanalyzed and error expressions
Original merge commit: 0978d17010d65560e54f18ed6cf972bc66f58820

######

Change ExprRef to return reference
ExprRef never returns a null pointer so change it to return a reference.
Add overloading for parser::Variable as they now also hold a typedExpr.
Change some of the functions that the result of ExprRef is passed to so
that they take references instead of pointers.


Original commit: 2cdab6b166aa71b273b64b1f6cf520d07637b534
-------

Distinguish between unanalyzed and error expressions
When an Expr or Variable is analyzed, always fill in a GenericExprWrapper
for it. That now holds the result of expression analysis which is
std::nullopt when there is an error.

After the ExprChecker pass has finished, the typedExpr data member of
Variables and top-level Exprs should be filled in. Assert that is the
case when we access them.


Original commit: 192a4e08554b105c48fcd65e6f8ed4860e64f544
-------