Add some example OMP stuff
Add CharBlock data member to OmpClause and OmpBlockDirective.
This is not necessarily the best place for them.

Define SHOW_ALL_SOURCE_MEMBERS in dump-parse-tree.h to see CharBlocks
in parse tree dump output.

In resolve-names.cc add OmpVisitor as a place to put OMP-related
stuff. Create a scope for OmpBlockDirective. When OmpObject is
encountered and contains a name, declare it as a local entity.
When it names a common block, check that it exists. None of this
is necessarily the right thing to do, it's just examples.