[flang-compiler/f18#509] Fix CHECK failure on undeclared type parameter
In the example below, the compiler gets an internal error when
processing the derived-type-spec for the declaration of `x`.
When the search for type parameter `k` fails it means an error
has already occurred and we should skip that parameter.

```
type :: t(k)
end type
type(t):: x
end
```

Original-commit: flang-compiler/f18@a4bf1922e5f2e4bc4d16a940bbb2367e171eb271
Reviewed-on: https://github.com/flang-compiler/f18/pull/509
Tree-same-pre-rewrite: true