Fixes for DO constraint C1124
Original merge commit: 00d74be74c7a822f53abe21e2bac89075af325de

######

Fixed the formatting of an "if" statement.


Original commit: 0cf71144529e9ebd625acc9a64b9318c890b669d
-------

Fixes for DO constraint C1124 The constraint says that a variable-name in a locality-spec shall be the name of a variable in the innermost executable construct or scoping unit that includes the DO CONCURRENT statement.  This check was already being made in resolve-names.cc for LOCAL and LOCAL_INIT locality specs but not for the SHARED locality spec.  Also, there was some code in check-do.cc that was intended to be used to enforce this constraint at some time in the future.
I added code to resolve-names.cc to extend the checking to the SHARED locality
spec and removed the unused code from check-do.cc.  I also extended the
existing tests in resolve35.f90 to exercise the new code.


Original commit: b421923d9403c3efe320e5ec2b722b88d1fdbd73
-------